Separate 'get_secrets' func from 'arcade_search.google tool'
