+++ toolkits/search/arcade_search/tools/google.py
@@ -1,9 +1,10 @@
 import json
-import os
-from typing import Annotated, Any, Optional
+from typing import Annotated
 
 import serpapi
 from arcade.sdk import tool
+
+from arcade_search.tools.utils import get_secret
 
 
 @tool
@@ -25,12 +26,3 @@ async def search_google(
     organic_results = results.get("organic_results", [])
 
     return json.dumps(organic_results[:n_results])
-
-
-def get_secret(name: str, default: Optional[Any] = None) -> Any:
-    secret = os.getenv(name)
-    if secret is None:
-        if default is not None:
-            return default
-        raise ValueError(f"Secret {name} is not set.")
-    return secret
